# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Меморіал API
  version: "1.0.0"
  summary: Публічний та партнерський API національної/регіональної платформи памʼяті захисників.
  description: |
    Контракт відповідає Master Plan v1.0 (розділ 18 «API») та ER-моделі docs/db/schema.dbml.
    Аутентифікація: OAuth2/OIDC (BankID/Дія.Підпис) для інтерактивних клієнтів,
    API-key (Bearer) для програмного доступу з rate-limit.
    Усі мутуючі запити підтримують заголовок `Idempotency-Key`.
    Пагінація — cursor-based (`cursor`, `limit`), максимум `limit=100`.
  contact:
    name: Платформа «Меморіал» — команда розробки
...
servers:
  - url: https://api.memorial.gov.ua/v1
    description: Production
  - url: https://api.staging.memorial.gov.ua/v1
    description: Staging
...
security:
  - OAuth2: []
  - ApiKeyAuth: []
paths:
...
  /places:
    get:
      tags: [Places]
      summary: Список місць памʼяті (для карти)
      parameters:
        - name: bbox
          in: query
          schema: { type: string, example: "30.1,46.3,30.9,46.6" }
          description: minLon,minLat,maxLon,maxLat
        - name: layer
          in: query
          schema:
            type: string
            enum: [battle, memorial, burial, monument, museum_site, frontline_segment]
        - name: date_from
          in: query
          schema: { type: string, format: date }
        - name: date_to
          in: query
          schema: { type: string, format: date }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: GeoJSON FeatureCollection
          content:
            application/geo+json:
              schema: { $ref: '#/components/schemas/GeoJSONFeatureCollection' }
  /places/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Places]
      summary: Картка місця
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Place' }
...
  /archive/search:
    get:
      tags: [Archive]
      summary: Розширений пошук по архіву (повнотекст OCR + фасети)
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: kind
          in: query
          schema:
            type: string
            enum: [photo, video, audio, document, document_3d, panorama]
        - name: date_from
          in: query
          schema: { type: string, format: date }
        - name: date_to
          in: query
          schema: { type: string, format: date }
        - name: rights_statement
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PagedResponse'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/MediaAssetSummary' }
  /media/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Archive]
      summary: Метадані медіаобʼєкта
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MediaAsset' }
...
  /moderation/queue:
    get:
      tags: [Moderation]
      summary: Черга модерації (UGC, заявки, правки)
      security: [{ OAuth2: [moderate:content] }]
      parameters:
        - name: entity_type
          in: query
          schema:
            type: string
            enum: [memory_ugc, defender_submission, content_revision]
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PagedResponse'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/ModerationItem' }
  /moderation/queue/{id}/decision:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Moderation]
      summary: Прийняти рішення (approve/reject)
      security: [{ OAuth2: [moderate:content] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [approve, reject] }
                reason: { type: string }
      responses:
        '200': { description: Рішення застосовано }

  # ---------------- SEARCH ----------------
...
components:
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.memorial.gov.ua/oauth2/authorize
          tokenUrl: https://auth.memorial.gov.ua/oauth2/token
          scopes:
            read:defenders: Читання профілів
            write:defenders: Редагування профілів
            write:ugc: Додавання спогадів/заявок
            write:submissions: Подання заявок
            partner:read: Читання даних партнера
            partner:write: Завантаження даних партнера
            moderate:content: Модерація черги
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
      description: Курсор пагінації з попередньої відповіді
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  schemas:
...
    PagedResponse:
      type: object
      properties:
        next_cursor: { type: string, nullable: true }
        total_estimate: { type: integer }
...
    GeoJSONFeatureCollection:
      type: object
      properties:
        type: { type: string, enum: [FeatureCollection] }
        features:
          type: array
          items:
            type: object
            properties:
              type: { type: string, enum: [Feature] }
              geometry: { type: object }
              properties:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  name: { type: string }
                  place_type: { type: string }
    Place:
      type: object
      properties:
        id: { type: string, format: uuid }
        type: { type: string }
        name: { type: string }
        description: { type: string }
        lon: { type: number }
        lat: { type: number }
        region_name: { type: string }
        cover_media_url: { type: string, nullable: true }
        related_defenders_count: { type: integer }
...
    MediaAssetSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        kind: { type: string }
        title: { type: string, nullable: true }
        thumbnail_url: { type: string, nullable: true }
        rights_statement: { type: string }
        date: { type: string, nullable: true }
    MediaAsset:
      allOf:
        - $ref: '#/components/schemas/MediaAssetSummary'
        - type: object
          properties:
            ocr_text: { type: string, nullable: true }
            exif: { type: object, nullable: true }
            provenance: { type: string, nullable: true }
            iiif_manifest_url: { type: string, nullable: true }
            derivatives:
              type: array
              items:
                type: object
                properties:
                  type: { type: string }
                  url: { type: string }
...
    ModerationItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        entity_type: { type: string }
        entity_id: { type: string, format: uuid }
        preview: { type: string }
        submitted_by: { type: string }
        ai_flags:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
...